pyproject.toml
[tool.poetry]
name = "atomic-habits"
version = "0.1.0"
description = ""
authors = ["Pinchuk Sergey <Pinboyer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
django-stubs = "^5.1.1"
black = "^24.10.0"
django-rest-framework = "^0.1.0"
python-dotenv = "^1.0.1"
drf-jwt = "^1.19.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
django-cors-headers = "^4.6.0"
drf-yasg = "^1.21.8"
flake8 = "^7.1.1"
requests = "^2.32.3"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
redis = "^5.2.0"
coverage = "^7.6.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
